Refactor: 게시판 카테고리 조회에 따라 id로 navbar 변경
