--- src/components/board/BoardCategory/style.ts
@@ -1,8 +1,0 @@
-import styled from "styled-components";
-
-const Category = styled.div`
-  display: flex;
-  flex-direction: column;
-`;
-
-export { Category };
